+++ SIOSR/Controllers/PenggalanganDanaController.cs
@@ -1,12 +1,9 @@
-using System;
-using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using SIOSR.Data;
-using SIOSR.Models.PenggalanganDana;
+using SIOSR.Models.App;
 
 namespace SIOSR.Controllers
 {
