+++ SIOSR/Views/Donasi/Create.cshtml
@@ -1,4 +1,4 @@
-@model SIOSR.Models.PenggalanganDana.Donasi
+@model SIOSR.Models.App.Donasi
 
 @{
     ViewData["Title"] = "Create";
