+++ SIOSR/Views/Donasi/Delete.cshtml
@@ -1,4 +1,4 @@
-@model SIOSR.Models.PenggalanganDana.Donasi
+@model SIOSR.Models.App.Donasi
 
 @{
     ViewData["Title"] = "Delete";
