+++ SIOSR/Views/Donasi/Index.cshtml
@@ -1,4 +1,4 @@
-@model IEnumerable<SIOSR.Models.PenggalanganDana.Donasi>
+@model IEnumerable<SIOSR.Models.App.Donasi>
 
 @{
     ViewData["Title"] = "Index";
