+++ SIOSR/Views/PenggalanganDana/Create.cshtml
@@ -1,4 +1,4 @@
-@model SIOSR.Models.PenggalanganDana.PenggalanganDana
+@model SIOSR.Models.App.PenggalanganDana
 
 @{
     ViewData["Title"] = "Create";
