+++ SIOSR/Views/PenggalanganDana/Edit.cshtml
@@ -1,4 +1,4 @@
-@model SIOSR.Models.PenggalanganDana.PenggalanganDana
+@model SIOSR.Models.App.PenggalanganDana
 
 @{
     ViewData["Title"] = "Edit";
