+++ SIOSR/Views/PenggalanganDana/Index.cshtml
@@ -1,4 +1,4 @@
-@model IEnumerable<SIOSR.Models.PenggalanganDana.PenggalanganDana>
+@model IEnumerable<SIOSR.Models.App.PenggalanganDana>
 
 @{
     ViewData["Title"] = "Index";
